feat: add $last property to table
